backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
openai>=1.6.1
pinecone-client>=3.1.0
youtube-transcript-api>=0.6.1
rank-bm25>=0.2.2
numpy>=1.26.2

# Blog Writing Agent (BWA)
langgraph>=0.2.0
tavily-python>=0.3.0
huggingface-hub>=0.20.0
httpx>=0.25.0

# Utilities
python-dotenv==1.0.0
certifi
